requirements.txt
numpy
pandas
yamale
mysql-connector-python
sqlalchemy

